GKB ASSESMENT by Sandeep Boddu/requirements.txt
mysql-connector-python
Flask
pillow
flask-cors
gevent 
gevent-websocket
Werkzeug

python -m io
